requirements.txt
python-telegram-bot==20.7
Pillow==10.2.0
asyncpg==0.29.0
APScheduler==3.10.4
python-dotenv==1.0.1
